Implement individual classes for each inverter domain
