--- kostal/const.py
@@ -1,100 +1,0 @@
-#!/usr/bin/env python3
-
-class ActualAnalogInputs:
-    ANALOG1 = 167772417
-    ANALOG2 = 167772673
-    ANALOG3 = 167772929
-    ANALOG4 = 167773185
-
-
-class ActualBattery:
-    VOLTAGE = 33556226
-    CHARGE = 33556229
-    CURRENT = 33556238
-    CURRENT_DIR = 33556230
-    CHARGE_CYCLES = 33556228
-    TEMPERATURE = 33556227
-
-
-class ActualGrid:
-    GRID_OUTPUT_POWER = 67109120
-    GRID_FREQ = 67110400
-    GRID_COS_PHI = 67110656
-    GRID_LIMITATION = 67110144
-    GRID_VOLTAGE_L1 = 67109378
-    GRID_CURRENT_L1 = 67109377
-    GRID_POWER_L1 = 67109379
-    GRID_VOLTAGE_L2 = 67109634
-    GRID_CURRENT_L2 = 67109633
-    GRID_POWER_L2 = 67109635
-    GRID_VOLTAGE_L3 = 67109890
-    GRID_CURRENT_L3 = 67109889
-    GRID_POWER_L3 = 67109891
-
-
-class ActualHome:
-    ACT_HOME_CONSUMPTION_SOLAR = 83886336
-    ACT_HOME_CONSUMPTION_BATTERY = 83886592
-    ACT_HOME_CONSUMPTION_GRID = 83886848
-    PHASE_SELECTIVE_CONSUMPTION_L1 = 83887106
-    PHASE_SELECTIVE_CONSUMPTION_L2 = 83887362
-    PHASE_SELECTIVE_CONSUMPTION_L3 = 83887618
-
-
-class ActualGenerator:
-    DC_1_VOLTAGE = 33555202
-    DC_1_CURRENT = 33555201
-    DC_1_POWER = 33555203
-    DC_2_VOLTAGE = 33555458
-    DC_2_CURRENT = 33555457
-    DC_2_POWER = 33555459
-    DC_3_VOLTAGE = 33555714
-    DC_3_CURRENT = 33555713
-    DC_3_POWER = 33555715
-
-
-class ActualSZeroIn:
-    S0_IN_PULSE_COUNT = 184549632
-    LOG_INTERVAL = 150995968
-
-
-class Home:
-    DC_POWER_PV = 33556736
-    AC_POWER = 67109120
-    OWN_CONSUMPTION = 83888128
-    BATTERY_STATE_OF_CHARGE = 33556229
-    # TODO: statusCodes
-    OPERATING_STATUS = 16780032
-
-
-class SettingsGeneral:
-    INVERTER_NAME = 16777984
-    INVERTER_MAKE = 16780544
-
-
-class InfoVersions:
-    VERSION_UI = 16779267
-    VERSION_FW = 16779265
-    VERSION_HW = 16779266
-    VERSION_PAR = 16779268
-    SERIAL_NUMBER = 16777728
-    ARTICLE_NUMBER = 16777472
-    COUNTRY_SETTINGS_NAME = 16779522
-    COUNTRY_SETTINGS_VERSION = 16779521
-
-
-class StatisticDay:
-    YIELD = 251658754
-    HOME_CONSUMPTION = 251659010
-    SELF_CONSUMPTION = 251659266
-    SELF_CONSUMPTION_RATE = 251659278
-    AUTONOMY_DEGREE = 251659279
-
-
-class StatisticTotal:
-    YIELD = 251658753
-    OPERATING_TIME = 251658496
-    HOME_CONSUMPTION = 251659009
-    SELF_CONSUMPTION = 251659265
-    SELF_CONSUMPTION_RATE = 251659280
-    AUTONOMY_DEGREE = 251659281
